Changed balance to be a double instead of a float, and used decimal formatting to correctly output to 2 decimal places when required.

+++ src/main/java/BankingApp/User/Transactions.java
@@ -3,6 +3,6 @@ package BankingApp.User;
 public interface Transactions {
 
     void seeBalance();
-    String depositMoney(float toDeposit);
-    String withdrawMoney(float toWithdraw);
+    String depositMoney(double toDeposit);
+    String withdrawMoney(double toWithdraw);
 }
